Re-write the tox hooks implementation for tox 3
This is the re-write of the existing code (existing code is still in
place in this commit) to enable the tox 3 hooks implementation to work
with the config style that also works with tox 4 hooks implementation.
There's quite a bit of code to resolve envs and do fallbacks that had to
be replicated as the tox 4 config relies on tox "doing the work",
whereas it's much trickier to get tox 3 to do it, due to how tox 3 hooks
require the addition of 'completed' testenv objects, whereas tox 4 can
be persuaded to use 'almost' tox.ini expressions.